chore: Move function that created the `FileNode` entries from uris to `~/data/folder/api.ts`
